Add middleware for user extraction

+++ part4/blogs/controllers/blogs.js
@@ -11,12 +11,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const newBlog = request.body
-
-  const decodedToken = jwt.verify(request.token, process.env.secret)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   if (!newBlog.likes) {
     newBlog.likes = 0
@@ -34,8 +29,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user._id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response
+      .status(400)
+      .json({ error: 'user invalid' })
+      .end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
